% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.home.bias.R
\name{add.home.bias}
\alias{add.home.bias}
\title{Add the effect of home-team bias to population picks}
\usage{
add.home.bias(teams, league = c("men", "women"), year = 2018)
}
\arguments{
\item{teams}{character vector of names of teams to whom home-team bias is to
be applied (must match name column of pred.pop.[league].[year])}

\item{league}{which league: "men" (default) or "women", for prob.source.}

\item{year}{year of tournament}
}
\value{
a 64-by-6 matrix matching population pick matrix from year, except
  that specified teams benefit from greater advancement probability
  attributable to home-team bias. Rows correspond to teams, and columns
  correspond to rounds, with the values giving the probability of each team
  winning in each round.
}
\description{
Add the effect of home-team bias to population picks
}
\details{
This function implements a rule of thumb gleamed from an article
  published at CBS Sports on March 10, 2016, by Brad Null of Bracket Voodoo,
  titled "Homer bias is real and it will derail your March Madness bracket".
  The sparse-detail results published there are consistent with adding a
  constant +3/4 to the log-odds of the conditional probability of the "home"
  team winning in each round. For more details, see
  https://github.com/elishayer/mRchmadness/issues/13. Null cautions in his
  piece that "schools with smaller undergraduate enrollment (Duke, Xavier,
  Villanova) tend to have a smaller regional footprint", so only use this
  adjustment if you have a good reason to think the picks in your pool would
  be biased (for example a pool of UNC alumni).
}
\examples{
add.home.bias(teams = 'UNC', league = 'men', year = 2018)
}
\author{
sspowers
}
