% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bracket.2017.women.R
\name{bracket.2017.women}
\alias{bracket.2017.women}
\title{2017 Women's March Madness bracket}
\format{character vector of length 64}
\description{
This dataset contains the ESPN team ids of the 64 teams in the
  2017 March Madness women's bracket. The teams are ordered by overall seed,
  such that the first four team ids correspond to the four #1 seeds.
}
