% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bracket.men.R
\name{bracket.men.2023}
\alias{bracket.men.2023}
\title{2023 Men's March Madness bracket}
\format{
character vector of length 64
}
\description{
This dataset contains the ESPN team ids of the 64 teams in the
  2023 March Madness men's bracket. The teams are ordered by overall seed,
  such that the first four team ids correspond to the four #1 seeds.
}
