% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bradley.terry.R
\name{bradley.terry}
\alias{bradley.terry}
\title{Fit a Bradley-Terry model on game score data}
\usage{
bradley.terry(games)
}
\arguments{
\item{games}{data.frame with the following columns: game.id, home.id,
away.id, home.score, away.score, neutral, ot (matched by output of
scrape.game.results)}
}
\description{
Fit a Bradley-Terry model on game score data
}
\examples{
  probability.matrix = bradley.terry(games.2017)
}
\author{
sspowers
}
