% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current.year.R
\docType{data}
\name{current.year}
\alias{current.year}
\title{Current year}
\format{An object of class \code{numeric} of length 1.}
\usage{
current.year
}
\description{
Your (almost) one stop shop for updating the current year
}
\keyword{datasets}
