% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold.R
\name{fold}
\alias{fold}
\title{Fold a vector onto itself}
\usage{
fold(x, block.size = 1)
}
\arguments{
\item{x}{a vector}

\item{block.size}{the size of groups in which to block the data}
}
\value{
a new vector in the following order: first block, last block,
  second block, second-to-last block, ...
}
\description{
Fold a vector onto itself
}
\examples{
fold(1:64)
}
\author{
sspowers
}
