% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games2017.R
\name{games.2017}
\alias{games.2017}
\title{2017 NCAA Men's Basketball Game-by-Game Results}
\format{data frame with 5848 rows and 7 variables}
\source{
ESPN
}
\description{
This dataset contains the game results of the
  college men's basketball 2016-2017 season, identifying
  the game, the home and away teams, (using the ESPN ID system),
  the home and away scores, the location at which the game was
  played, and whether the game went into overtime. Teams that
  are not in Division I are uniformly identified by the string
  'NA' as their id.
}
