% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games.men.2018.R
\name{games.men.2018}
\alias{games.men.2018}
\title{2018 NCAA Men's Basketball Game-by-Game Results}
\format{data frame with 5352 rows and 7 variables}
\source{
ESPN
}
\description{
This dataset contains the game results of the
  college men's basketball 2017-2018 season, identifying
  the game, the home and away teams, (using the ESPN ID system),
  the home and away scores, whether the game was played at a
  neutral arena, and whether the game went into overtime. Teams
  that are not in Division I are uniformly identified by the
  string 'NA' as their id.
  The dataset is currently updated only through February 2018.
}
