% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.bracket.filled.R
\name{input.bracket.filled}
\alias{input.bracket.filled}
\title{Reorder from pool-order to the internal order}
\usage{
input.bracket.filled(bracket.picks)
}
\arguments{
\item{bracket.picks}{a length-63 character vector giving the 63 game
outcomes in the order from reading off a standard-formatted bracket. More
specifically, this is the first 32 first-round games going by region from
South to West to East to Midwest, and within region the winner of the game
between, in order, 1-16, 8-9, 5-12, 4-13, 6-11, 3-14, 7-10, 2-15. Then the
winners of the second round in the same region order, and the winner of
1-16-8-9, 5-12-4-13, 6-11-3-14, and 7-10-2-15. At this point the pattern
should be clear. See https://github.com/elishayer/mRchmadness/issues/23
for more details.}
}
\value{
a vector giving the picked bracket in the way used elsewhere
  in mRchmadness (see sim.bracket documentation)
}
\description{
Reorder from pool-order to the internal order
}
\author{
elishayer
}
