% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.population.team.names.R
\name{map.population.team.names}
\alias{map.population.team.names}
\title{Map team names from population pick distribution}
\usage{
map.population.team.names(pop.names, teams, teams.map.538)
}
\arguments{
\item{pop.names}{a vector of team names scraped from the population pick distribution
(should be length 68)}

\item{teams}{data.frame from mRchmadness::teams.men or mRchmadness::teams.women}

\item{teams.map.538}{data.frame of tournament team IDs and names from 538,
used as the target for mapping}

\item{return}{a list of two data.frames: full.map and auto.map. The former is the full map for
all of the teams, and the latter shows which ones were generated by fuzzy string matching.}
}
\description{
This mapper has not been exhaustively tested
}
\author{
saberpowers
}
