% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mRchmadness.R
\docType{package}
\name{mRchmadness}
\alias{mRchmadness}
\alias{mRchmadness-package}
\title{March Madness Bracket Package}
\description{
\code{mRchmadness} provides utilities to gather NCAA Men's
Basketball data, predict the win probability of matchups,
and produce a bracket optimized to a specified criterion
}
