% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.538.men.R
\name{pred.538.men.2017}
\alias{pred.538.men.2017}
\title{2017 March Madness 538 Win Probability Predictions}
\format{
data frame with 64 rows and 7 variables
}
\source{
\url{https://projects.fivethirtyeight.com/2017-march-madness-predictions/}
}
\description{
This dataset contains the win probabilities forecasted
  by the website FiveThirtyEight immediately after the First Four
  and before any other games had been played in the March Madness
  tournament of the 2016-2017 season. The columns are the names of
  the teams, and FiveThirtyEight's projected probability of reaching each round
  of the tournament.
}
