% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.538.men.2018.R
\name{pred.538.men.2018}
\alias{pred.538.men.2018}
\title{2018 March Madness 538 Win Probability Predictions}
\format{data frame with 68 rows and 7 variables}
\source{
\url{https://projects.fivethirtyeight.com/2018-march-madness-predictions/}
}
\description{
This dataset contains the win probabilities forecasted
  by the website FiveThirtyEight before any games have been played
  in the March Madness tournament of the 2017-2018 season. The columns
  are the names of the teams, and FiveThirtyEight's projected
  probability of reaching each round of the tournament.
}
