% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.538.women.R
\name{pred.538.women.2022}
\alias{pred.538.women.2022}
\title{2022 Women's March Madness 538 Win Probability Predictions}
\format{
data frame with 64 rows and 7 variables
}
\source{
\url{https://projects.fivethirtyeight.com/2022-march-madness-predictions/}
}
\description{
This dataset contains the win probabilities forecasted by the
  website FiveThirtyEight for the Women's March Madness tournament of the
  2021-2022 season. The columns are the names of the teams, and
  FiveThirtyEight's projected probability of reaching each round.
}
