% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.kenpom.men.2018.R
\name{pred.kenpom.men.2018}
\alias{pred.kenpom.men.2018}
\title{2018 March Madness KenPom Win Probability Predictions}
\format{data frame with 64 rows and 7 variables}
\source{
\url{https://kenpom.com/blog/2018-ncaa-tournament-probabilities/}
}
\description{
This dataset contains the win probabilities forecasted
  by the website KenPom before any games have been played
  in the March Madness tournament of the 2017-2018 season. The columns
  are the names of the teams, and KenPom's projected
  probability of reaching each round of the tournament.
}
