% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.pop.women.2016.R
\name{pred.pop.women.2016}
\alias{pred.pop.women.2016}
\title{2016 March Madness Women's Population Pick Distribution}
\format{data frame with 64 rows and 7 variables}
\source{
\url{http://games.espn.com/tournament-challenge-bracket-women/2016/en/whopickedwhom}
}
\description{
This dataset contains the percent of brackets submitted
  to ESPN that include each of the 64 teams in Women's March Madness 2016
  reaching and winning in each successive round of the tournament.
}
