% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.app.R
\name{run.app}
\alias{run.app}
\title{Run the Shiny app allowing for interaction with the bracket
production given user-entered criteria}
\usage{
run.app()
}
\description{
Run the Shiny app allowing for interaction with the bracket
production given user-entered criteria
}
\author{
eshayer
}
