% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.game.results.R
\name{scrape.game.results}
\alias{scrape.game.results}
\title{Scrape the game-by-game results of the NCAA MBB seaon}
\usage{
scrape.game.results(year)
}
\description{
Scrape the game-by-game results of the NCAA MBB seaon
}
\examples{
game.results = scrape.game.results(2017)
}
\author{
eshayer
}
