% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.game.results.R
\name{scrape.team.game.results}
\alias{scrape.team.game.results}
\title{Scrape game results for a single team-year combination}
\usage{
scrape.team.game.results(year, team.id, league)
}
\arguments{
\item{year}{a character value representing a year}

\item{team.id}{an ESPN team id}

\item{league}{either 'mens' or 'womens'}
}
\value{
data.frame of game data for the team-year
}
\description{
Scrape game results for a single team-year combination
}
\author{
eshayer
}
