% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.game.results.R
\name{scrape.teams}
\alias{scrape.teams}
\title{Scrape the team names and ids from the ESPN NCAA MBB index}
\usage{
scrape.teams()
}
\description{
Scrape the team names and ids from the ESPN NCAA MBB index
}
\author{
eshayer
}
