% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.bracket.matrix.R
\name{sim.bracket.matrix}
\alias{sim.bracket.matrix}
\title{Simulate the full bracket starting with an empty bracket}
\usage{
sim.bracket.matrix(
  prob.matrix,
  league,
  num.reps,
  outcome,
  round,
  teams.remaining,
  untangling.indices
)
}
\arguments{
\item{prob.matrix}{a matrix of probabilities, with rows and columns
corresponding to teams, matching the output of bradley.terry()}

\item{league}{which league: "men" (default) or "women"}

\item{num.reps}{number of simulations to perform}

\item{outcome}{passed in from sim.bracket()}

\item{round}{passed in from sim.bracket()}

\item{teams.remaining}{passed in from sim.bracket()}

\item{untangling.indices}{passed in from sim.bracket()}
}
\value{
a 63-by-num.reps matrix storing the simulation outcome, each
  column encoding the outcome for a single simulation in the following
  order: seeds 1 through 32 after round 1, seeds 1 through 16 after round 2,
  seeds 1 through 8 after round 3, seeds 1 through 4 after round 4,
  seeds 1 and 2 after round 5, and finally seed 1 after round 6 (the
  champion)
}
\description{
Simulate the full bracket starting with an empty bracket
}
\author{
sspowers
}
