% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.bracket.source.R
\name{sim.bracket.source}
\alias{sim.bracket.source}
\title{Simulate the full bracket starting with an empty bracket}
\usage{
sim.bracket.source(bracket.empty, prob.source, year, num.reps, outcome, round,
  teams, untangling.indices)
}
\arguments{
\item{bracket.empty}{a length-64 character vector giving the field of 64
teams in the tournament, in order of initial overall seeding}

\item{prob.source}{source from which to use round probabilities for
simulation --- "pop": ESPN's population of picks,
"Pom": Ken Pomeroy's predictions (kenpom.com), or
"538": predictions form fivethirtyeight.com.}

\item{year}{year of tournament, used for prob.source.}

\item{num.reps}{number of simulations to perform}

\item{outcome}{passed in from sim.bracket()}

\item{round}{passed in from sim.bracket()}

\item{teams}{passed in from sim.bracket()}

\item{untangling.indices}{passed in from sim.bracket()}
}
\value{
a 63-by-num.reps matrix storing the simulation outcome, each
  column encoding the outcome for a single simulation in the following
  order: seeds 1 through 32 after round 1, seeds 1 through 16 after round 2,
  seeds 1 through 8 after round 3, seeds 1 through 4 after round 4,
  seeds 1 and 2 after round 5, and finally seed 1 after round 6 (the
  champion)
}
\description{
Simulate the full bracket starting with an empty bracket
}
\author{
sspowers
}
