% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.men.R
\name{teams.men}
\alias{teams.men}
\title{NCAA Men's Basketball Teams}
\format{
data frame with 351 rows and 5 variables
}
\source{
\url{http://www.espn.com/mens-college-basketball/teams}
}
\description{
This dataset the names of each team in NCAA Men's
  Basketball from ESPN, their ESPN team id, their name as it
  appears on Who Picked Whom, FiveThirtyEight, and KenPom.
  The last three name fields are only available for those teams
  who played in March Madness in 2016 to 2021, and are otherwise
  NA.
}
