% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\name{teams}
\alias{teams}
\title{NCAA Men's Basketball Teams}
\format{data frame with 351 rows and 4 variables}
\source{
\url{http://www.espn.com/mens-college-basketball/teams}
}
\description{
This dataset the names of each team in NCAA Men's
  Basketball from ESPN, their ESPN team id, their name as it
  appears on the Who Picked Whom, and the FiveThirtyEight name.
  The last two name fields are only available for those teams
  who played in March Madness in 2016 or 2017, and are otherwise
  NA.
}
