% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.women.R
\name{teams.women}
\alias{teams.women}
\title{NCAA Women's Basketball Teams}
\format{data frame with 413 rows and 4 variables}
\source{
\url{http://www.espn.com/womens-college-basketball/teams}
}
\description{
This dataset the names of each team in NCAA Women's
  Basketball from ESPN and their ESPN team id
}
