% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfold.R
\name{unfold}
\alias{unfold}
\title{Unfold a vector (the inverse of the fold function)}
\usage{
unfold(x, block.size = 1)
}
\arguments{
\item{x}{a vector}

\item{block.size}{the size of groups in which to block the data}
}
\value{
a vector in the following order: block 1, block 3, ..., block n-1,
  block n, block n-2, ..., block 2.
}
\description{
Unfold a vector (the inverse of the fold function)
}
\author{
sspowers
}
