% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bradley.terry.R
\name{bradley.terry}
\alias{bradley.terry}
\title{Fit a Bradley-Terry model on game score data}
\usage{
bradley.terry(games)
}
\arguments{
\item{games}{data.frame with the following columns: game.id, home.id,
away.id, home.score, away.score, neutral, ot (matched by output of
scrape.game.results)}
}
\value{
matrix of win probabilities, with rows and columns labeled by team.
  Each entry gives the probability of the team corresponding to that row
  beating the team corresponding to that column.
}
\description{
Fit a Bradley-Terry model on game score data
}
\examples{
  prob = bradley.terry(games = games.2017)
}
\author{
sspowers
}
