% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.bracket.R
\name{draw.bracket}
\alias{draw.bracket}
\title{Plot bracket to device}
\usage{
draw.bracket(bracket.empty, bracket.filled = NULL)
}
\arguments{
\item{bracket.empty}{a length-64 character vector giving the field of 64
teams in the tournament, in order of initial overall seeding}

\item{bracket.filled}{an optional length-63 character vector encoding
tournament results (matching output from simulate.bracket)}
}
\description{
Plot bracket to device
}
\examples{
probability.matrix = bradley.terry(games.2017)
picks = sim.bracket(bracket.2017, probability.matrix)
draw.bracket(bracket.2017, picks)
}
\author{
sspowers
}
