% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mRchmadness.R
\docType{package}
\name{mRchmadness}
\alias{mRchmadness-package}
\alias{mRchmadness}
\title{March Madness Bracket Package}
\description{
\code{mRchmadness} provides utilities to gather NCAA Men's
Basketball data, predict the win probability of matchups,
and produce a bracket optimized to a specified criterion
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/elishayer/mRchmadness}
}

}
\author{
\strong{Maintainer}: Eli Shayer \email{eshayer@stanford.edu}

Authors:
\itemize{
  \item Scott Powers \email{saberpowers@gmail.com}
}

}
