% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.game.results.R
\name{scrape.game.results}
\alias{scrape.game.results}
\title{Scrape the game-by-game results of the NCAA MBB seaon}
\usage{
scrape.game.results(year)
}
\arguments{
\item{year}{a numeric value of the year, between 2002 and 2017 inclusive}
}
\value{
data.frame with game-by-game results
}
\description{
Scrape the game-by-game results of the NCAA MBB seaon
}
\author{
eshayer
}
