% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.game.results.R
\name{scrape.team.game.results}
\alias{scrape.team.game.results}
\title{Scrape game results for a single team-year combination}
\usage{
scrape.team.game.results(year, id)
}
\description{
Scrape game results for a single team-year combination
}
\examples{
stanford.2017.games = scrape.team.game.results(2017, 24)
}
