% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.bracket.R
\name{sim.bracket}
\alias{sim.bracket}
\title{Simulate the full bracket starting with an empty bracket}
\usage{
sim.bracket(bracket.empty, probability.matrix, num.reps = 1)
}
\arguments{
\item{bracket.empty}{a length-64 character vector giving the field of 64
teams in the tournament, in order of initial overall seeding}

\item{probability.matrix}{a matrix of probabilities, with rows and columns
corresponding to teams, matching the output of bradley.terry()}

\item{num.reps}{number of simulations to perform (default is 1)}
}
\description{
Simulate the full bracket starting with an empty bracket
}
\examples{
probability.matrix = bradley.terry(games.2017)
sim.bracket(bracket.2017, probability.matrix)
}
\author{
sspowers
}
