% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.bracket.source.R
\name{sim.bracket.source}
\alias{sim.bracket.source}
\title{Simulate the full bracket starting with an empty bracket}
\usage{
sim.bracket.source(prob.source, league, year, home.teams, num.reps, outcome,
  round, teams.remaining, untangling.indices)
}
\arguments{
\item{prob.source}{source from which to use round probabilities for
simulation --- "pop": ESPN's population of picks,
"kenpom": Ken Pomeroy's predictions (kenpom.com), or
"538": predictions form fivethirtyeight.com.}

\item{league}{which league: "men" (default) or "women", for prob.source.}

\item{year}{year of tournament, used for prob.source.}

\item{home.teams}{character vector of names of teams to whom home-team bias
is to be applied (must match name column of pred.pop.[league].[year]).
Ignored unless prob.source is "pop" (see ?add.home.bias for details).}

\item{num.reps}{number of simulations to perform}

\item{outcome}{passed in from sim.bracket()}

\item{round}{passed in from sim.bracket()}

\item{teams.remaining}{passed in from sim.bracket()}

\item{untangling.indices}{passed in from sim.bracket()}
}
\value{
a 63-by-num.reps matrix storing the simulation outcome, each
  column encoding the outcome for a single simulation in the following
  order: seeds 1 through 32 after round 1, seeds 1 through 16 after round 2,
  seeds 1 through 8 after round 3, seeds 1 through 4 after round 4,
  seeds 1 and 2 after round 5, and finally seed 1 after round 6 (the
  champion)
}
\description{
Simulate the full bracket starting with an empty bracket
}
\author{
sspowers
}
