% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.bracket.R
\name{test.bracket}
\alias{test.bracket}
\title{Test a bracket}
\usage{
test.bracket(bracket.empty, probability.matrix, bracket.picks, pool.size = 30,
  num.sims = 1000, bonus.round = c(1, 2, 4, 8, 16, 32),
  bonus.seed = rep(0, 16), bonus.combine = c("add", "multiply"))
}
\arguments{
\item{bracket.empty}{a length-64 character vector giving the field of 64
teams in the tournament, in order of initial overall seeding}

\item{probability.matrix}{a matrix of probabilities, with rows and columns
corresponding to teams, matching the output of bradley.terry()}

\item{bracket.picks}{an length-63 character vector encoding your picks
(this is the bracket to be evaluated)}

\item{pool.size}{number of brackets in your pool (excluding yours), matters
only if criterion == "win" (default is 30)}

\item{num.sims}{number of simulations over which to evaluate the candidate
brackets (default is 1000)}

\item{bonus.round}{a length-6 vector giving the number of points awarded in
your pool's scoring rules for correct picks in each round (default is
2^round)}

\item{bonus.seed}{a length-16 vector giving the bonus awarded for correctly
picking winner based on winner's seed (default is zero)}

\item{bonus.combine}{how to combine the round bonus with the seed bonus to
get the number of points awarded for each correct pick: "add" (default) or
multiply}
}
\description{
Test a bracket
}
\examples{
probability.matrix = bradley.terry(games.2017)
my.bracket = find.bracket(bracket.2017, probability.matrix)
result = test.bracket(bracket.2017, probability.matrix, my.bracket)
}
\author{
sspowers
}
