% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.bracket.R
\name{test.bracket}
\alias{test.bracket}
\title{Test a bracket}
\usage{
test.bracket(bracket.empty, bracket.picks, prob.matrix = NULL,
  prob.source = c("pop", "Pom", "538"), pool.source = c("pop", "Pom",
  "538"), league = c("men", "women"), year = 2017, pool.size = 30,
  num.sims = 1000, bonus.round = c(1, 2, 4, 8, 16, 32),
  bonus.seed = rep(0, 16), bonus.combine = c("add", "multiply"),
  print.progress = TRUE, shiny.progress = NULL)
}
\arguments{
\item{bracket.empty}{a length-64 character vector giving the field of 64
teams in the tournament, in order of initial overall seeding}

\item{bracket.picks}{an length-63 character vector encoding your picks
(this is the bracket to be evaluated)}

\item{prob.matrix}{a matrix of probabilities, with rows and columns
corresponding to teams, matching the output of bradley.terry().
This probabilities are used to simulate outcomes on
which to evaluate bracket.picks. If NULL, prob.source is used.}

\item{prob.source}{source from which to use round probabilities to simulate
outcomes --- "pop": ESPN's population of picks (default),
"Pom": Ken Pomeroy's predictions (kenpom.com), or
"538": predictions form fivethirtyeight.com.
Ignored if prob.matrix is specified.}

\item{pool.source}{source from which to use round probabilities to simulate
entries of opponents in pool. Same options as prob.source.}

\item{league}{which league: "men" (default) or "women", for pool.source.}

\item{year}{year of tournament, used for prob.source.
Ignored if prob.matrix is specified.}

\item{pool.size}{number of brackets in your pool (excluding yours), matters
only if criterion == "win" (default is 30)}

\item{num.sims}{number of simulations over which to evaluate the candidate
brackets (default is 1000)}

\item{bonus.round}{a length-6 vector giving the number of points awarded in
your pool's scoring rules for correct picks in each round (default is
2^round)}

\item{bonus.seed}{a length-16 vector giving the bonus awarded for correctly
picking winner based on winner's seed (default is zero)}

\item{bonus.combine}{how to combine the round bonus with the seed bonus to
get the number of points awarded for each correct pick: "add" (default) or
multiply}

\item{print.progress}{should progress be printed to console?}

\item{shiny.progress}{a shiny::Progress object used only for the Shiny app}
}
\description{
Test a bracket
}
\examples{
prob.matrix = bradley.terry(games = games.men.2017)
my.bracket = find.bracket(bracket.empty = bracket.men.2017,
  prob.matrix = prob.matrix, pool.source = "pop", league = "men",
  year = 2017)
result = test.bracket(bracket.empty = bracket.men.2017,
  bracket.picks = my.bracket, prob.matrix = prob.matrix,
  pool.source = "pop", league = "men", year = 2017)
}
\author{
sspowers
}
